git clone + user interruption safety

+++ virttest/asset.py
@@ -307,9 +307,10 @@ def download_test_provider(provider, update=False):
         ref = provider_info.get('ref')
         pubkey = provider_info.get('pubkey')
         download_dst = data_dir.get_test_provider_dir(provider)
+        dir_existed = os.path.exists(download_dst)
+        repo_downloaded = os.path.isdir(os.path.join(download_dst, '.git'))
+        original_dir = os.getcwd()
         try:
-            repo_downloaded = os.path.isdir(os.path.join(download_dst, '.git'))
-            original_dir = os.getcwd()
             if not repo_downloaded or update:
                 download_dst = git.get_repo(uri=uri, branch=branch, commit=ref,
                                             destination_dir=download_dst)
@@ -319,15 +320,23 @@ def download_test_provider(provider, update=False):
                 except process.CmdError:
                     pass
                 process.run('git pull origin %s' % branch)
-            os.chdir(download_dst)
-            process.system('git log -1')
-            os.chdir(original_dir)
-        except (KeyboardInterrupt, process.CmdError):
-            logging.error('Cleaning up provider %s download dir %s', provider,
-                          download_dst)
-            if os.path.isdir(download_dst):
+        except:
+            if not dir_existed and os.path.isdir(download_dst):
+                logging.error('Cleaning up provider %s download dir %s', provider,
+                              download_dst)
                 shutil.rmtree(download_dst)
             raise
+
+        # sanity check to ensure the git repository is OK:
+        try:
+            os.chdir(download_dst)
+            process.system('git log -1')
+        except process.CmdError:
+            logging.error('Something is unexpectedly wrong with the git repository at %s',
+                          download_dst)
+            raise
+        finally:
+            os.chdir(original_dir)
 
 
 def download_all_test_providers(update=False):
